# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/client/wallet": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "check the balance for client",
                "description": "check the balance for client",
                "operationId": "check the balance for client",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/client/wallet/charge": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "charge the balance for client",
                "description": "charge the balance for client",
                "operationId": "charge the balance for client",
                "parameters": [
                    {
                        "name": "card_number",
                        "in": "path",
                        "description": "card number (for credit)",
                        "required": true
                    },
                    {
                        "name": "exp_month",
                        "in": "path",
                        "description": "expiration month (for credit)",
                        "required": true
                    },
                    {
                        "name": "exp_year",
                        "in": "path",
                        "description": "expiration year (for credit)",
                        "required": true
                    },
                    {
                        "name": "CVC",
                        "in": "path",
                        "description": "CVC (for credit)",
                        "required": true
                    },
                    {
                        "name": "amount",
                        "in": "path",
                        "description": "amount (for credit)",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/client/trips": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "get all trips for client",
                "description": "get all trips for client",
                "operationId": "get all trips for client",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/client/trips/joined": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "get joined trips for client",
                "description": "get joined trips for client",
                "operationId": "get joined trips for client",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/client/trips/{id}": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "join a trip",
                "description": "join a trip",
                "operationId": "join a trip",
                "parameters": [
                    {
                        "name": "payment_method",
                        "in": "path",
                        "description": "payment method (wallet or credit)",
                        "required": true
                    },
                    {
                        "name": "card_number",
                        "in": "path",
                        "description": "card number (for credit)",
                        "required": true
                    },
                    {
                        "name": "exp_month",
                        "in": "path",
                        "description": "expiration month (for credit)",
                        "required": true
                    },
                    {
                        "name": "exp_year",
                        "in": "path",
                        "description": "expiration year (for credit)",
                        "required": true
                    },
                    {
                        "name": "CVC",
                        "in": "path",
                        "description": "CVC (for credit)",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/client/trips/cancel/{id}": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "cancel a trip",
                "description": "cancel a trip",
                "operationId": "cancel a trip",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/client/complaints": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "make a complaint for client",
                "description": "make a complaint for client",
                "operationId": "make a complaint for client",
                "parameters": [
                    {
                        "name": "message",
                        "in": "path",
                        "description": "message",
                        "required": true
                    },
                    {
                        "name": "trip_id",
                        "in": "path",
                        "description": "trip id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
        "/api/organization/wallet": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "check the balance",
                "description": "check the balance",
                "operationId": "check the balance",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}